jakairta_backend/requirements.txt
django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-environ==0.11.2
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
Pillow==10.4.0
requests==2.32.3
APScheduler==3.10.4
django-apscheduler==0.6.2
pyfcm==2.0.7
gunicorn==22.0.0
whitenoise==6.7.0
